+++ js/data.js
@@ -2,12 +2,17 @@
 import {getRandomElement} from './util.js';
 import {getRandomElementsArray} from './util.js'
 import {getRandomNumber} from './util.js';
-import {Card} from './Card.js';
+import {getRandomFloatNumber} from './util.js'
 
 const typePremises = ['palace', 'flat', 'house', 'bungalow'];
 const elementsEntryDeparture = ['12:00', '13:00', '14:00'];
 const facilities = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
 const listPhotos = ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+const COORDINATES_FROMX = 35.65000;
+const COORDINATES_BEFOREX = 35.70000;
+const COORDINATES_FROMY = 139.70000;
+const COORDINATES_BEFOREY = 139.80000;
+const COORDINATES_AFTERPOINT = 5;
 
 const getInfoAd = (coordinatesAxisX, coordinatesAxisY) => {
   return {
@@ -31,6 +36,17 @@ const createAuthor = () => {
   };
 };
 
+const Card = function () {
+  const coordinateX = getRandomFloatNumber(COORDINATES_FROMX, COORDINATES_BEFOREX, COORDINATES_AFTERPOINT);
+  const coordinateY = getRandomFloatNumber(COORDINATES_FROMY, COORDINATES_BEFOREY, COORDINATES_AFTERPOINT);
+  this.author = createAuthor();
+  this.offer = getInfoAd(coordinateX, coordinateY);
+  this.location = {
+    x: coordinateX,
+    y: coordinateY,
+  };
+};
+
 const getCardsArray = (quantityElements) => {
   return new Array(quantityElements).fill(null).map(() => new Card());
 };
@@ -38,3 +54,4 @@ const getCardsArray = (quantityElements) => {
 export {createAuthor};
 export {getCardsArray};
 export {getInfoAd};
+export {Card};
